requirements.txt
Flask
imutils
Keras
opencv-python
Pillow
tensorflow
scikit-image
pandas
numpy
-e .